+++ actix-http/CHANGES.md
@@ -16,6 +16,8 @@
 * `impl Display` for `header::Quality`. [#2486]
 * Connection data set through the `on_connect_ext` callbacks is now accessible only from the new `Request::conn_data()` method. [#2491]
 * `Request::take_conn_data()`. [#2491]
+* `Request::take_req_data()`. [#2487]
+* `impl Clone` for `RequestHead`. [#2487]
 
 ### Changed
 * Rename `body::BoxBody::{from_body => new}`. [#2468]
@@ -40,6 +42,7 @@
 [#2468]: https://github.com/actix/actix-web/pull/2468
 [#1920]: https://github.com/actix/actix-web/pull/1920
 [#2486]: https://github.com/actix/actix-web/pull/2486
+[#2487]: https://github.com/actix/actix-web/pull/2487
 [#2488]: https://github.com/actix/actix-web/pull/2488
 [#2491]: https://github.com/actix/actix-web/pull/2491